refactor(validation): split the 1,462-line module behind a re-exporting facade
Closes #906

43 files under functions/ import validation.js, so the split goes behind a
facade rather than by rewriting call sites. NOT ONE call site outside
functions/utils/ changes in this PR — that is what keeps a 1,400-line move
reviewable.

validation.js is now 45 lines of re-exports. Implementation moved to
functions/utils/validation/:

  fieldLimits.js   57   FIELD_LIMITS
  identity.js     105   email, password, UUID, role
  strings.js       60   sanitizeString, sanitizeOptionalText, requiredFields
  urls.js         280   normalizeHttpUrl, the safeReflect* family, sanitizers
  contact.js       69   phone, postal code
  datetime.js     279   ISO date, time, set times, performance date, doors JSON
  ids.js           68   validateId, validateIdArray
  schema.js       561   validateEntity, VALIDATION_SCHEMAS
  response.js      23   validationErrorResponse

schema.js sits downstream of everything, since validateEntity and
VALIDATION_SCHEMAS call into nearly every other module. Three internal-only
exports exist to serve it without re-declaring constants — VALID_ROLES from
identity.js, POSTAL_CODE_REGEX and PHONE_REGEX from contact.js. They are not on
the public facade. No cycles.

Implemented by a delegated Sonnet run. Verified independently before commit
rather than on the report:

- Extracted all 29 exported function bodies from the pre-split file and from the
  new modules, normalised whitespace, compared: 29/29 byte-identical. The move
  is mechanical, with no behaviour change hidden in it.
- The two const exports cannot be diffed as text, so compared them as VALUES —
  imported FIELD_LIMITS and VALIDATION_SCHEMAS before and after, serialised with
  regexes stringified: identical.
- Confirmed the importer count is still 43.
- Re-ran make gate: exit 0, 1234 backend / 1094 frontend.

The surface guard committed in the previous commit (33 assertions over all 31
export names and their types) passes throughout, which is what proves the facade
is complete: a route importing a dropped symbol would fail at request time, not
at import time, so a green unit run alone would not have caught it.

Worth recording, from the delegate's own report: a ` ` escape in
normalizePostalCode's regex was briefly transcoded into a literal NBSP byte
while being copied. Runtime-equivalent, but it would have left an invisible
character in the source. Verified fixed — the escape is intact and no raw NBSP
bytes exist in any new module.
