+++ functions/utils/validation.js
@@ -1,1462 +1,45 @@
-// Input validation utilities for Cloudflare Workers
-// Provides reusable validation functions for API endpoints
-
-import { validate as validateEmail } from "email-validator";
-
-/**
- * UUID validation regex (RFC 4122 compliant)
- * Validates UUID v1-5 format
- */
-const UUID_REGEX = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
-
-/**
- * ISO 8601 date format regex
- * Matches YYYY-MM-DD and full ISO datetime formats
- */
-const ISO_DATE_REGEX = /^\d{4}-\d{2}-\d{2}(T\d{2}:\d{2}:\d{2}(\.\d{3})?(Z|[+-]\d{2}:\d{2})?)?$/;
-
-/**
- * Postal code regex (supports US ZIP and Canadian postal codes)
- */
-const POSTAL_CODE_REGEX = /^(?:\d{5}(?:-\d{4})?|[A-Za-z]\d[A-Za-z]\s*\d[A-Za-z]\d)$/;
-
-/**
- * Phone number regex (permissive, digits + formatting)
- */
-const PHONE_REGEX = /^\+?[\d\s().-]{7,20}$/;
-
-/**
- * Valid user roles in the system
- * Update this array when adding new roles
- */
-const VALID_ROLES = ["admin", "editor", "viewer"];
-
-/**
- * Valid event statuses
- */
-const VALID_EVENT_STATUSES = ["draft", "published", "archived"];
-const ALLOWED_EXTERNAL_PROTOCOLS = new Set(["http:", "https:"]);
-
-/**
- * 24-hour "HH:MM" time regex for `events.doors_json` values (#569).
- */
-const DOORS_TIME_REGEX = /^([01]\d|2[0-3]):[0-5]\d$/;
-
-/**
- * Control characters to remove during sanitization
- * Removes: null bytes (\x00), control characters except tab/newline (\x0B-\x1F), and DEL (\x7F)
- */
-// eslint-disable-next-line no-control-regex -- intentional: matches control chars for security sanitization
-const CONTROL_CHARS_REGEX = /[\x00-\x08\x0B-\x0C\x0E-\x1F\x7F]/g;
-
-/**
- * Field length limits - centralized configuration
- */
-export const FIELD_LIMITS = {
-  // User fields
-  email: { min: 5, max: 255 },
-  password: { min: 12, max: 128 },
-  userName: { min: 2, max: 100 },
-  userFirstName: { min: 1, max: 60 },
-  userLastName: { min: 1, max: 60 },
-
-  // Venue fields
-  venueName: { min: 1, max: 200 },
-  venueAddress: { min: 0, max: 200 },
-  venueAddressLine1: { min: 0, max: 200 },
-  venueAddressLine2: { min: 0, max: 200 },
-  venueCity: { min: 0, max: 100 },
-  venueRegion: { min: 0, max: 100 },
-  venuePostal: { min: 0, max: 20 },
-  venueCountry: { min: 0, max: 100 },
-  venuePhone: { min: 0, max: 25 },
-  venueContactEmail: { min: 0, max: 255 },
-
-  // Band fields
-  bandName: { min: 1, max: 200 },
-  bandOrigin: { min: 0, max: 100 },
-  bandOriginCity: { min: 0, max: 100 },
-  bandOriginRegion: { min: 0, max: 100 },
-  bandGenre: { min: 0, max: 100 },
-  bandDescription: { min: 0, max: 5000 },
-  bandUrl: { min: 0, max: 500 },
-  socialHandle: { min: 0, max: 100 },
-  bandContactEmail: { min: 0, max: 255 },
-
-  // Performance fields
-  performanceNotes: { min: 0, max: 1000 },
-
-  // Event fields
-  eventName: { min: 3, max: 200 },
-  eventSlug: { min: 3, max: 100 },
-  ticketLink: { min: 0, max: 500 },
-  eventPosterUrl: { min: 0, max: 500 },
-  eventDescription: { min: 0, max: 5000 },
-  eventCity: { min: 0, max: 100 },
-  eventVenueInfo: { min: 0, max: 5000 },
-  eventSocialLinks: { min: 0, max: 2000 },
-  eventThemeColors: { min: 0, max: 1000 },
-  eventDoorsJson: { min: 0, max: 2000 },
-
-  // Generic
-  url: { min: 0, max: 2000 },
-  shortText: { min: 0, max: 255 },
-  longText: { min: 0, max: 10000 },
-};
-
-/**
- * Validate email format
- * @param {string} email - Email address to validate
- * @returns {boolean} True if valid email format
- */
-export function isValidEmail(email) {
-  if (!email || typeof email !== "string") {
-    return false;
-  }
-  return validateEmail(email.trim());
-}
-
-/**
- * Validate password strength
- * @param {string} password - Password to validate
- * @param {Object} options - Validation options
- * @param {number} options.minLength - Minimum length (default: 8)
- * @param {boolean} options.requireUppercase - Require uppercase letter (default: false)
- * @param {boolean} options.requireLowercase - Require lowercase letter (default: false)
- * @param {boolean} options.requireNumber - Require number (default: false)
- * @param {boolean} options.requireSpecial - Require special character (default: false)
- * @returns {Object} { valid: boolean, errors: string[] }
- */
-export function validatePassword(password, options = {}) {
-  const {
-    minLength = FIELD_LIMITS.password.min,
-    requireUppercase = false,
-    requireLowercase = false,
-    requireNumber = false,
-    requireSpecial = false,
-  } = options;
-
-  const errors = [];
-
-  if (!password || typeof password !== "string") {
-    errors.push("Password is required");
-    return { valid: false, errors };
-  }
-
-  if (password.length < minLength) {
-    errors.push(`Password must be at least ${minLength} characters`);
-  }
-
-  if (requireUppercase && !/[A-Z]/.test(password)) {
-    errors.push("Password must contain at least one uppercase letter");
-  }
-
-  if (requireLowercase && !/[a-z]/.test(password)) {
-    errors.push("Password must contain at least one lowercase letter");
-  }
-
-  if (requireNumber && !/\d/.test(password)) {
-    errors.push("Password must contain at least one number");
-  }
-
-  // Matches any character that is NOT alphanumeric (letters or numbers)
-  // This includes symbols, punctuation, and special characters
-  if (requireSpecial && !/[^A-Za-z0-9]/.test(password)) {
-    errors.push("Password must contain at least one special character");
-  }
-
-  return { valid: errors.length === 0, errors };
-}
-
-/**
- * Validate required fields in an object
- * @param {Object} data - Data object to validate
- * @param {string[]} requiredFields - Array of required field names
- * @returns {Object} { valid: boolean, missing: string[] }
- */
-export function validateRequiredFields(data, requiredFields) {
-  if (!data || typeof data !== "object") {
-    return { valid: false, missing: requiredFields };
-  }
-
-  const missing = requiredFields.filter((field) => {
-    const value = data[field];
-    return value === undefined || value === null || value === "";
-  });
-
-  return { valid: missing.length === 0, missing };
-}
-
-/**
- * Validate UUID format
- * @param {string} uuid - UUID to validate
- * @returns {boolean} True if valid UUID format
- */
-export function isValidUUID(uuid) {
-  if (!uuid || typeof uuid !== "string") {
-    return false;
-  }
-  return UUID_REGEX.test(uuid);
-}
-
-/**
- * Validate role value
- * @param {string} role - Role to validate
- * @returns {boolean} True if valid role
- */
-export function isValidRole(role) {
-  if (!role || typeof role !== "string") {
-    return false;
-  }
-  return VALID_ROLES.includes(role);
-}
-
-/**
- * Sanitize string input by removing dangerous characters
- * Note: This is a basic sanitization. For HTML content, use DOMPurify on client-side
- * @param {string} input - Input to sanitize
- * @returns {string} Sanitized input
- */
-export function sanitizeString(input) {
-  if (!input || typeof input !== "string") {
-    return "";
-  }
-
-  // Remove null bytes (\x00), control characters except tab/newline (\x0B-\x1F), and DEL (\x7F)
-  return input.replace(CONTROL_CHARS_REGEX, "").trim();
-}
-
-export function sanitizeOptionalText(value, maxLength, label) {
-  if (value === undefined || value === null) {
-    return null;
-  }
-
-  const text = sanitizeString(String(value));
-  if (!text) {
-    return null;
-  }
-
-  if (maxLength !== undefined && text.length > maxLength) {
-    throw new Error(`${label} must be no more than ${maxLength} characters`);
-  }
-
-  return text;
-}
-
-function sanitizeOptionalHandle(value, maxLength, label) {
-  const text = sanitizeOptionalText(value, maxLength, label);
-  if (!text) {
-    return null;
-  }
-
-  if (/\s/.test(text)) {
-    throw new Error(`${label} must not contain spaces`);
-  }
-
-  // A handle is a bare @name — it must never carry a URL scheme (javascript:,
-  // data:, etc.) or path separators. Legitimate Instagram/X/TikTok handles
-  // never contain these characters, so any occurrence means the field is
-  // being used to smuggle a URL/scheme past the handle-only contract.
-  if (/[:/\\]/.test(text)) {
-    throw new Error(`${label} must not contain a URL scheme or path separator`);
-  }
-
-  return text;
-}
-
-function parseJsonInput(value, label) {
-  if (value === undefined || value === null || value === "") {
-    return null;
-  }
-
-  if (typeof value === "string") {
-    try {
-      return JSON.parse(value);
-    } catch {
-      throw new Error(`${label} must be valid JSON`);
-    }
-  }
-
-  if (typeof value === "object") {
-    return value;
-  }
-
-  throw new Error(`${label} must be valid JSON`);
-}
-
-function sanitizeOptionalHandleOrUrl(value, maxLength, label) {
-  if (value === undefined || value === null || value === "") {
-    return null;
-  }
-
-  const text = sanitizeString(String(value));
-  if (!text) {
-    return null;
-  }
-
-  if (/^https?:\/\//i.test(text)) {
-    return sanitizeOptionalHttpUrl(text, maxLength, label);
-  }
-
-  return sanitizeOptionalHandle(text, maxLength, label);
-}
-
-export function normalizeHttpUrl(url) {
-  if (!url || typeof url !== "string") {
-    return null;
-  }
-
-  const trimmedUrl = url.trim();
-  if (!trimmedUrl) {
-    return null;
-  }
-
-  try {
-    const parsed = new URL(trimmedUrl);
-    if (!ALLOWED_EXTERNAL_PROTOCOLS.has(parsed.protocol)) {
-      return null;
-    }
-    return parsed.toString();
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Sanitize a stored handle-or-URL value for safe reflection in an API
- * response. This is a read-path counterpart to `sanitizeOptionalHandleOrUrl`
- * (the write-path validator) — it exists because rows written before the
- * write-path guard existed (or written by a bypassed/legacy path) may still
- * contain unsafe scheme values like `javascript:alert(1)` in the DB. Never
- * reflect those verbatim.
- *
- * - Trimmed first, matching the write-path sanitizers, so a legacy value
- *   with stray whitespace (" https://example.com") is recovered as a URL
- *   instead of falling to the handle branch and being dropped. Trimming
- *   cannot launder a scheme: a trimmed "javascript:…" still carries the
- *   colon and is rejected below.
- * - http(s):// values are re-validated/normalized via `normalizeHttpUrl`.
- * - Anything else is treated as a bare handle. A colon is the necessary
- *   condition for any URL scheme (javascript:, data:, vbscript:, …), so a
- *   handle containing one is rejected. Deliberately lenient on slashes so
- *   quirky legacy handles like "instagram.com/band" aren't dropped.
- *
- * @param {*} value - Raw stored value
- * @returns {string|null} Safe value to reflect, or null
- */
-export function safeReflectHandleOrUrl(value) {
-  if (typeof value !== "string") {
-    return null;
-  }
-
-  const text = value.trim();
-  if (!text) {
-    return null;
-  }
-
-  if (/^https?:\/\//i.test(text)) {
-    return normalizeHttpUrl(text);
-  }
-
-  return text.includes(":") ? null : text;
-}
-
-/**
- * Sanitize a stored social-links JSON string for safe reflection in an API
- * response. Parses the JSON defensively (malformed JSON → `{}`) and routes
- * each field through the appropriate read-path sanitizer: handle fields via
- * `safeReflectHandleOrUrl`, everything else via `normalizeHttpUrl`.
- *
- * @param {string} jsonString - Raw `social_links` column value
- * @param {string[]} handleFields - Keys that hold handles rather than URLs
- * @returns {Object} Plain object of sanitized values
- */
-export function safeReflectSocialLinks(jsonString, handleFields = ["instagram"]) {
-  let parsed;
-  try {
-    parsed = JSON.parse(jsonString);
-  } catch {
-    return {};
-  }
-
-  if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) {
-    return {};
-  }
-
-  const sanitized = {};
-  for (const [key, value] of Object.entries(parsed)) {
-    sanitized[key] = handleFields.includes(key) ? safeReflectHandleOrUrl(value) : normalizeHttpUrl(value);
-  }
-
-  return sanitized;
-}
-
-/**
- * String-in/string-out wrapper around `safeReflectSocialLinks`, for admin
- * read endpoints (#493) that reflect `social_links` as a raw JSON string —
- * the admin frontend (RosterTab.jsx, EventFormModal.jsx, LineupTab.jsx,
- * admin/utils/pickerFormData.js) parses that string itself, so the response
- * shape must stay a string rather than switching to a parsed object.
- *
- * A `null`/`undefined` column value (no social links set) is passed through
- * unchanged rather than coerced to `"{}"`, so callers that never had a
- * `social_links` value don't see one appear in the response.
- *
- * @param {string|null|undefined} jsonString - Raw `social_links` column value
- * @param {string[]} handleFields - Keys that hold handles rather than URLs
- * @returns {string|null|undefined} Sanitized JSON string, or the original
- *   nullish value
- */
-export function safeReflectSocialLinksString(jsonString, handleFields = ["instagram"]) {
-  if (jsonString === null || jsonString === undefined) {
-    return jsonString;
-  }
-  return JSON.stringify(safeReflectSocialLinks(jsonString, handleFields));
-}
-
-export function sanitizeOptionalHttpUrl(value, maxLength = FIELD_LIMITS.url.max, label = "URL") {
-  const text = sanitizeOptionalText(value, maxLength, label);
-  if (!text) {
-    return null;
-  }
-
-  const normalized = normalizeHttpUrl(text);
-  if (!normalized) {
-    throw new Error(`${label} must start with http:// or https://`);
-  }
-
-  return normalized;
-}
-
-export function sanitizeBandSocialLinks(value) {
-  const parsed = parseJsonInput(value, "Social links");
-  if (!parsed) {
-    return null;
-  }
-
-  if (Array.isArray(parsed) || typeof parsed !== "object") {
-    throw new Error("Social links must be a JSON object");
-  }
-
-  const sanitized = {
-    website: sanitizeOptionalHttpUrl(parsed.website, FIELD_LIMITS.bandUrl.max, "Website URL"),
-    instagram: sanitizeOptionalHandleOrUrl(parsed.instagram, FIELD_LIMITS.socialHandle.max, "Instagram"),
-    bandcamp: sanitizeOptionalHttpUrl(parsed.bandcamp, FIELD_LIMITS.bandUrl.max, "Bandcamp URL"),
-    facebook: sanitizeOptionalHttpUrl(parsed.facebook, FIELD_LIMITS.bandUrl.max, "Facebook URL"),
-    youtube: sanitizeOptionalHttpUrl(parsed.youtube, FIELD_LIMITS.bandUrl.max, "YouTube URL"),
-    spotify: sanitizeOptionalHttpUrl(parsed.spotify, FIELD_LIMITS.bandUrl.max, "Spotify URL"),
-    apple_music: sanitizeOptionalHttpUrl(parsed.apple_music, FIELD_LIMITS.bandUrl.max, "Apple Music URL"),
-    linktree: sanitizeOptionalHttpUrl(parsed.linktree, FIELD_LIMITS.bandUrl.max, "Linktree URL"),
-  };
-
-  return Object.values(sanitized).some(Boolean) ? JSON.stringify(sanitized) : null;
-}
-
-export function sanitizeEventSocialLinks(value) {
-  const parsed = parseJsonInput(value, "Social links");
-  if (!parsed) {
-    return null;
-  }
-
-  if (Array.isArray(parsed) || typeof parsed !== "object") {
-    throw new Error("Social links must be a JSON object");
-  }
-
-  const sanitized = {
-    website: sanitizeOptionalHttpUrl(parsed.website, FIELD_LIMITS.ticketLink.max, "Website URL"),
-    instagram: sanitizeOptionalHandleOrUrl(parsed.instagram, FIELD_LIMITS.ticketLink.max, "Instagram"),
-    facebook: sanitizeOptionalHttpUrl(parsed.facebook, FIELD_LIMITS.ticketLink.max, "Facebook URL"),
-    x: sanitizeOptionalHandleOrUrl(parsed.x, FIELD_LIMITS.ticketLink.max, "X / Twitter"),
-    tiktok: sanitizeOptionalHandleOrUrl(parsed.tiktok, FIELD_LIMITS.ticketLink.max, "TikTok"),
-    youtube: sanitizeOptionalHttpUrl(parsed.youtube, FIELD_LIMITS.ticketLink.max, "YouTube URL"),
-    bandcamp: sanitizeOptionalHttpUrl(parsed.bandcamp, FIELD_LIMITS.ticketLink.max, "Bandcamp URL"),
-  };
-
-  return Object.values(sanitized).some(Boolean) ? JSON.stringify(sanitized) : null;
-}
-
-export function sanitizeVenueInfo(value) {
-  const parsed = parseJsonInput(value, "Venue info");
-  if (!parsed) {
-    return null;
-  }
-
-  if (!Array.isArray(parsed)) {
-    throw new Error("Venue info must be a JSON array");
-  }
-
-  const sanitizedVenues = parsed.map((venue, index) => {
-    if (!venue || typeof venue !== "object" || Array.isArray(venue)) {
-      throw new Error(`Venue ${index + 1} must be a JSON object`);
-    }
-
-    const name = sanitizeOptionalText(venue.name, FIELD_LIMITS.venueName.max, `Venue ${index + 1} name`);
-    if (!name) {
-      throw new Error(`Venue ${index + 1} name is required`);
-    }
-
-    return {
-      name,
-      address: sanitizeOptionalText(venue.address, FIELD_LIMITS.venueAddress.max, `Venue ${index + 1} address`),
-      note: sanitizeOptionalText(venue.note, FIELD_LIMITS.shortText.max, `Venue ${index + 1} note`),
-      googleMaps: sanitizeOptionalHttpUrl(venue.googleMaps, FIELD_LIMITS.url.max, `Venue ${index + 1} map link`),
-    };
-  });
-
-  return sanitizedVenues.length > 0 ? JSON.stringify(sanitizedVenues) : null;
-}
-
-/**
- * Validate URL format
- * @param {string} url - URL to validate
- * @returns {boolean} True if valid URL
- */
-export function isValidURL(url) {
-  return Boolean(normalizeHttpUrl(url));
-}
-
-/**
- * Validate phone number format (permissive)
- * @param {string} phone - Phone number to validate
- * @returns {boolean} True if valid or empty
- */
-export function isValidPhone(phone) {
-  if (!phone || typeof phone !== "string") {
-    return true;
-  }
-  return PHONE_REGEX.test(phone.trim());
-}
-
-/**
- * Validate postal code format (US ZIP or Canadian postal)
- * @param {string} postalCode - Postal code to validate
- * @returns {boolean} True if valid or empty
- */
-export function isValidPostalCode(postalCode) {
-  if (!postalCode || typeof postalCode !== "string") {
-    return true;
-  }
-  return POSTAL_CODE_REGEX.test(postalCode.trim());
-}
-
-/**
- * Normalize postal code to standard format
- * Canadian: "A1A 1A1" (uppercase, single space)
- * US: "12345" or "12345-6789" (trimmed)
- * @param {string} postalCode - Raw postal code
- * @returns {string|null} Normalized postal code or null if empty
- */
-export function normalizePostalCode(postalCode) {
-  if (!postalCode || typeof postalCode !== "string") {
-    return null;
-  }
-
-  const trimmed = postalCode.trim().toUpperCase();
-  if (!trimmed) return null;
-
-  // Check for Canadian format (A1A 1A1 or A1A1A1)
-  // Remove all whitespace/separators to check the raw alphanumeric sequence
-  const clean = trimmed.replace(/[\s\u00A0-]+/g, ""); // Remove spaces, nbsp, hyphens usually not in CA code
-
-  // Canadian: 6 chars, Letter-Digit-Letter Digit-Letter-Digit
-  if (clean.length === 6 && /^[A-Z]\d[A-Z]\d[A-Z]\d$/.test(clean)) {
-    return `${clean.slice(0, 3)} ${clean.slice(3)}`;
-  }
-
-  // If not identified as Canadian, return the trimmed original (handles US zip)
-  return trimmed;
-}
-
-/**
- * Validate date string (ISO 8601 format)
- * Strictly validates ISO 8601 format: YYYY-MM-DD or YYYY-MM-DDTHH:mm:ss.sssZ
- * @param {string} dateString - Date string to validate
- * @returns {boolean} True if valid ISO date format
- */
-export function isValidISODate(dateString) {
-  if (!dateString || typeof dateString !== "string") {
-    return false;
-  }
-
-  // First check if format matches ISO 8601
-  if (!ISO_DATE_REGEX.test(dateString)) {
-    return false;
-  }
-
-  // Then verify it's a valid date that can be parsed
-  const date = new Date(dateString);
-  return !isNaN(date.getTime());
-}
-
-/**
- * Create a validation error response
- * @param {string} message - Error message
- * @param {Object} details - Additional error details
- * @param {number} status - HTTP status code (default: 400)
- * @returns {Response} Error response
- */
-export function validationErrorResponse(message, details = {}, status = 400) {
-  return new Response(
-    JSON.stringify({
-      error: "Validation error",
-      message,
-      ...details,
-    }),
-    {
-      status,
-      headers: { "Content-Type": "application/json" },
-    },
-  );
-}
-
-/**
- * Validate time format (HH:MM) and logical validity
- * @param {string} time - Time string to validate
- * @returns {Object} { valid: boolean, error: string|null }
- */
-export function isValidTime(time) {
-  if (!time || typeof time !== "string") {
-    return { valid: false, error: "Time is required" };
-  }
-
-  // Check format
-  if (!/^\d{2}:\d{2}$/.test(time)) {
-    return { valid: false, error: "Time must be in HH:MM format" };
-  }
-
-  // Check logical validity
-  const [hours, minutes] = time.split(":").map(Number);
-  if (hours < 0 || hours > 23) {
-    return { valid: false, error: "Hours must be between 00 and 23" };
-  }
-  if (minutes < 0 || minutes > 59) {
-    return { valid: false, error: "Minutes must be between 00 and 59" };
-  }
-
-  return { valid: true, error: undefined };
-}
-
-/**
- * Reject a zero-length set — a performance whose end time equals its start.
- *
- * This is the canonical home for that rule. It existed inline, worded
- * differently, on two of the four write paths that accept a user-supplied
- * start AND end time, and not at all on the other two (`bands/import.js`, and
- * the PATCH in `bands/[id].js` — the path an operator actually uses to move a
- * set time during a show).
- *
- * The rule matters because a zero-length set is read two incompatible ways.
- * `normalizeEndMinutes` in `functions/utils/timeConflicts.js` treated
- * `end <= start` as a midnight crossing and added 24 hours, making the set
- * conflict with everything at its venue; the frontend's copy in
- * `frontend/src/admin/utils/timeUtils.js` used `end < start`, leaving a
- * zero-width interval that `intervalsOverlap` (strict `<`) matches against
- * nothing. Server and admin UI therefore disagreed about the same row.
- *
- * Rejecting the input is the fix rather than picking a side: a band plays
- * neither 0 minutes nor 24 hours, so both readings were wrong. With the input
- * impossible, the two comparisons cannot disagree — they are aligned anyway so
- * they cannot drift apart again.
- *
- * `end < start` stays legal: that is a real after-midnight set (23:30–00:30),
- * and 5 exist in production. A missing `end_time` also stays legal — 175 rows
- * have one, so the rule applies only when BOTH values are present.
- *
- * @param {string|null|undefined} startTime - HH:MM, or absent
- * @param {string|null|undefined} endTime - HH:MM, or absent
- * @returns {{valid: boolean, error: string|null}}
- */
-export function validateSetTimes(startTime, endTime) {
-  if (!startTime || !endTime) {
-    return { valid: true, error: undefined };
-  }
-  if (startTime === endTime) {
-    return { valid: false, error: "Start and end time cannot be the same" };
-  }
-  return { valid: true, error: undefined };
-}
-
-/**
- * Validate date string and check if it's a valid calendar date
- * @param {string} dateString - Date string in YYYY-MM-DD format
- * @returns {Object} { valid: boolean, error: string|null, date: Date|null }
- */
-export function validateDate(dateString) {
-  if (!dateString || typeof dateString !== "string") {
-    return { valid: false, error: "Date is required", date: null };
-  }
-
-  // Check format
-  if (!/^\d{4}-\d{2}-\d{2}$/.test(dateString)) {
-    return {
-      valid: false,
-      error: "Date must be in YYYY-MM-DD format",
-      date: null,
-    };
-  }
-
-  // Parse and validate the date
-  const [year, month, day] = dateString.split("-").map(Number);
-
-  // Check year range (reasonable bounds)
-  if (year < 2020 || year > 2100) {
-    return { valid: false, error: "Year must be between 2020 and 2100", date: null };
-  }
-
-  // Check month
-  if (month < 1 || month > 12) {
-    return { valid: false, error: "Month must be between 01 and 12", date: null };
-  }
-
-  // Check day (accounting for month lengths and leap years)
-  const daysInMonth = new Date(year, month, 0).getDate();
-  if (day < 1 || day > daysInMonth) {
-    return {
-      valid: false,
-      error: `Day must be between 01 and ${daysInMonth} for this month`,
-      date: null,
-    };
-  }
-
-  const date = new Date(dateString);
-  return { valid: true, error: undefined, date };
-}
-
-/**
- * Validate a performance's festival-day assignment (#540, multi-day events).
- *
- * `performanceDate` is optional — null/undefined/"" all mean "inherit the
- * event's single date" (the pre-#540 default for every existing row) and are
- * always valid. When provided, it must be a real YYYY-MM-DD calendar date
- * that falls within the event's festival-day span [event.date, event.end_date].
- * A null `event.end_date` collapses that span to the single `event.date`
- * (single-day events only ever have one valid value).
- *
- * Comparisons are plain lexicographic string comparisons on YYYY-MM-DD —
- * never `new Date(...)`, which parses bare date strings as UTC and drifts a
- * day in negative-offset timezones (see CLAUDE.md).
- *
- * @param {string|null|undefined} performanceDate
- * @param {{ date: string, end_date: string|null }|null} event
- * @returns {{ valid: boolean, error: string|null, value: string|null }}
- */
-export function validatePerformanceDate(performanceDate, event) {
-  if (performanceDate === undefined || performanceDate === null || performanceDate === "") {
-    return { valid: true, error: undefined, value: null };
-  }
-
-  if (typeof performanceDate !== "string") {
-    return { valid: false, error: "performance_date must be a YYYY-MM-DD string", value: null };
-  }
-
-  const dateCheck = validateDate(performanceDate);
-  if (!dateCheck.valid) {
-    return { valid: false, error: dateCheck.error, value: null };
-  }
-
-  const minDate = event?.date || null;
-  const maxDate = event?.end_date || event?.date || null;
-
-  if (!minDate || performanceDate < minDate || performanceDate > maxDate) {
-    return {
-      valid: false,
-      error: `performance_date must be between ${minDate || "the event start"} and ${maxDate || "the event end"}`,
-      value: null,
-    };
-  }
-
-  return { valid: true, error: undefined, value: performanceDate };
-}
-
-/**
- * Validate an event's per-day doors/gates-open time map (#569).
- *
- * `doorsJson` is optional — null/undefined/"" all mean "no doors info" (the
- * default for every pre-#569 row) and are always valid. When provided it
- * must be a JSON object whose keys are YYYY-MM-DD dates that fall within the
- * event's festival-day span [event.date, event.end_date] (same rule and same
- * lexicographic YYYY-MM-DD comparisons as `validatePerformanceDate` — never
- * `new Date(...)`, see CLAUDE.md) and whose values are 24-hour "HH:MM"
- * times. An empty object is normalized to null (equivalent to absent).
- *
- * @param {string|object|null|undefined} doorsJson - raw JSON string or already-parsed object
- * @param {{ date: string, end_date: string|null }|null} event
- * @returns {{ valid: boolean, error: string|null, value: string|null }}
- */
-export function validateDoorsJson(doorsJson, event) {
-  if (doorsJson === undefined || doorsJson === null || doorsJson === "") {
-    return { valid: true, error: undefined, value: null };
-  }
-
-  let parsed;
-  if (typeof doorsJson === "string") {
-    if (doorsJson.length > FIELD_LIMITS.eventDoorsJson.max) {
-      return {
-        valid: false,
-        error: `Doors times must be no more than ${FIELD_LIMITS.eventDoorsJson.max} characters`,
-        value: null,
-      };
-    }
-    try {
-      parsed = JSON.parse(doorsJson);
-    } catch {
-      return { valid: false, error: "Doors times must be valid JSON", value: null };
-    }
-  } else if (typeof doorsJson === "object") {
-    parsed = doorsJson;
-  } else {
-    return { valid: false, error: "Doors times must be a JSON object", value: null };
-  }
-
-  if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) {
-    return { valid: false, error: "Doors times must be a JSON object", value: null };
-  }
-
-  const entries = Object.entries(parsed);
-  if (entries.length === 0) {
-    return { valid: true, error: undefined, value: null };
-  }
-
-  const minDate = event?.date || null;
-  const maxDate = event?.end_date || event?.date || null;
-
-  for (const [dateKey, timeValue] of entries) {
-    const dateCheck = validateDate(dateKey);
-    if (!dateCheck.valid) {
-      return { valid: false, error: `Doors times key "${dateKey}" must be a valid YYYY-MM-DD date`, value: null };
-    }
-
-    if (!minDate || dateKey < minDate || dateKey > maxDate) {
-      return {
-        valid: false,
-        error: `Doors times date "${dateKey}" must be between ${minDate || "the event start"} and ${maxDate || "the event end"}`,
-        value: null,
-      };
-    }
-
-    if (typeof timeValue !== "string" || !DOORS_TIME_REGEX.test(timeValue)) {
-      return { valid: false, error: `Doors time for "${dateKey}" must be in 24-hour HH:MM format`, value: null };
-    }
-  }
-
-  const serialized = JSON.stringify(parsed);
-  if (serialized.length > FIELD_LIMITS.eventDoorsJson.max) {
-    return {
-      valid: false,
-      error: `Doors times must be no more than ${FIELD_LIMITS.eventDoorsJson.max} characters`,
-      value: null,
-    };
-  }
-
-  return { valid: true, error: undefined, value: serialized };
-}
-
-/**
- * Validate a positive integer ID
- * @param {any} id - ID to validate
- * @returns {Object} { valid: boolean, value: number|null, error: string|null }
- */
-export function validateId(id) {
-  if (id === undefined || id === null || id === "") {
-    return { valid: false, value: null, error: "ID is required" };
-  }
-
-  const numId = Number(id);
-  if (!Number.isInteger(numId) || numId < 1) {
-    return { valid: false, value: null, error: "ID must be a positive integer" };
-  }
-
-  return { valid: true, value: numId, error: undefined };
-}
-
-/**
- * Validate an array of IDs
- * @param {any} ids - Array of IDs to validate
- * @param {Object} options - Validation options
- * @param {number} options.maxLength - Maximum array length (default: 100)
- * @returns {Object} { valid: boolean, values: number[]|null, error: string|null }
- */
-export function validateIdArray(ids, options = {}) {
-  const { maxLength = 100 } = options;
-
-  if (!Array.isArray(ids)) {
-    return { valid: false, values: null, error: "Must be an array" };
-  }
-
-  if (ids.length === 0) {
-    return { valid: false, values: null, error: "Array cannot be empty" };
-  }
-
-  if (ids.length > maxLength) {
-    return {
-      valid: false,
-      values: null,
-      error: `Array cannot contain more than ${maxLength} items`,
-    };
-  }
-
-  const values = [];
-  for (let i = 0; i < ids.length; i++) {
-    const result = validateId(ids[i]);
-    if (!result.valid) {
-      return {
-        valid: false,
-        values: null,
-        error: `Invalid ID at index ${i}: ${result.error}`,
-      };
-    }
-    values.push(result.value);
-  }
-
-  // Check for duplicates
-  const uniqueValues = new Set(values);
-  if (uniqueValues.size !== values.length) {
-    return { valid: false, values: null, error: "Array contains duplicate IDs" };
-  }
-
-  return { valid: true, values, error: undefined };
-}
-
-/**
- * Comprehensive entity validation
- * @param {Object} data - Data to validate
- * @param {Object} schema - Validation schema
- * @returns {Object} { valid: boolean, errors: Object, sanitized: Object }
- */
-export function validateEntity(data, schema) {
-  const errors = {};
-  const sanitized = {};
-
-  for (const [field, rules] of Object.entries(schema)) {
-    const value = data[field];
-
-    // Check required
-    if (rules.required && (value === undefined || value === null || value === "")) {
-      errors[field] = `${rules.label || field} is required`;
-      continue;
-    }
-
-    // Skip validation for optional empty fields
-    if (!rules.required && (value === undefined || value === null || value === "")) {
-      sanitized[field] = rules.default !== undefined ? rules.default : null;
-      continue;
-    }
-
-    // Type validation
-    if (rules.type === "string") {
-      if (typeof value !== "string") {
-        errors[field] = `${rules.label || field} must be a string`;
-        continue;
-      }
-
-      const sanitizedValue = sanitizeString(value);
-
-      // Length validation
-      if (rules.min !== undefined && sanitizedValue.length < rules.min) {
-        errors[field] = `${rules.label || field} must be at least ${rules.min} characters`;
-        continue;
-      }
-      if (rules.max !== undefined && sanitizedValue.length > rules.max) {
-        errors[field] = `${rules.label || field} must be no more than ${rules.max} characters`;
-        continue;
-      }
-
-      // Pattern validation
-      if (rules.pattern && !rules.pattern.test(sanitizedValue)) {
-        errors[field] = rules.patternError || `${rules.label || field} has an invalid format`;
-        continue;
-      }
-
-      // Custom validation
-      if (rules.validate) {
-        const result = rules.validate(sanitizedValue);
-        if (!result.valid) {
-          errors[field] = result.error;
-          continue;
-        }
-      }
-
-      sanitized[field] = sanitizedValue;
-    } else if (rules.type === "email") {
-      if (typeof value !== "string" || !isValidEmail(value)) {
-        errors[field] = `${rules.label || field} must be a valid email address`;
-        continue;
-      }
-      const sanitizedEmail = value.trim().toLowerCase();
-      if (rules.min !== undefined && sanitizedEmail.length < rules.min) {
-        errors[field] = `${rules.label || field} must be at least ${rules.min} characters`;
-        continue;
-      }
-      if (rules.max !== undefined && sanitizedEmail.length > rules.max) {
-        errors[field] = `${rules.label || field} must be no more than ${rules.max} characters`;
-        continue;
-      }
-      sanitized[field] = sanitizedEmail;
-    } else if (rules.type === "url") {
-      if (typeof value !== "string") {
-        errors[field] = `${rules.label || field} must be a string`;
-        continue;
-      }
-      const trimmedUrl = value.trim();
-      if (trimmedUrl && !isValidURL(trimmedUrl)) {
-        errors[field] = `${rules.label || field} must be a valid URL`;
-        continue;
-      }
-      if (rules.max && trimmedUrl.length > rules.max) {
-        errors[field] = `${rules.label || field} must be no more than ${rules.max} characters`;
-        continue;
-      }
-      sanitized[field] = trimmedUrl || null;
-    } else if (rules.type === "time") {
-      const timeResult = isValidTime(value);
-      if (!timeResult.valid) {
-        errors[field] = timeResult.error;
-        continue;
-      }
-      sanitized[field] = value;
-    } else if (rules.type === "date") {
-      const dateResult = validateDate(value);
-      if (!dateResult.valid) {
-        errors[field] = dateResult.error;
-        continue;
-      }
-      sanitized[field] = value;
-    } else if (rules.type === "id") {
-      const idResult = validateId(value);
-      if (!idResult.valid) {
-        errors[field] = idResult.error;
-        continue;
-      }
-      sanitized[field] = idResult.value;
-    } else if (rules.type === "enum") {
-      if (!rules.values.includes(value)) {
-        errors[field] = `${rules.label || field} must be one of: ${rules.values.join(", ")}`;
-        continue;
-      }
-      sanitized[field] = value;
-    } else if (rules.type === "boolean") {
-      sanitized[field] = Boolean(value);
-    } else if (rules.type === "number") {
-      const numValue = Number(value);
-      if (isNaN(numValue)) {
-        errors[field] = `${rules.label || field} must be a number`;
-        continue;
-      }
-      if (rules.min !== undefined && numValue < rules.min) {
-        errors[field] = `${rules.label || field} must be at least ${rules.min}`;
-        continue;
-      }
-      if (rules.max !== undefined && numValue > rules.max) {
-        errors[field] = `${rules.label || field} must be no more than ${rules.max}`;
-        continue;
-      }
-      sanitized[field] = numValue;
-    }
-  }
-
-  return {
-    valid: Object.keys(errors).length === 0,
-    errors,
-    sanitized,
-  };
-}
-
-// Pre-defined validation schemas for common entities
-export const VALIDATION_SCHEMAS = {
-  venue: {
-    name: {
-      type: "string",
-      required: true,
-      label: "Venue name",
-      min: FIELD_LIMITS.venueName.min,
-      max: FIELD_LIMITS.venueName.max,
-    },
-    address: {
-      type: "string",
-      required: false,
-      label: "Address",
-      min: FIELD_LIMITS.venueAddress.min,
-      max: FIELD_LIMITS.venueAddress.max,
-      default: null,
-    },
-    address_line1: {
-      type: "string",
-      required: false,
-      label: "Street address",
-      min: FIELD_LIMITS.venueAddressLine1.min,
-      max: FIELD_LIMITS.venueAddressLine1.max,
-      default: null,
-    },
-    address_line2: {
-      type: "string",
-      required: false,
-      label: "Address line 2",
-      min: FIELD_LIMITS.venueAddressLine2.min,
-      max: FIELD_LIMITS.venueAddressLine2.max,
-      default: null,
-    },
-    city: {
-      type: "string",
-      required: false,
-      label: "City",
-      min: FIELD_LIMITS.venueCity.min,
-      max: FIELD_LIMITS.venueCity.max,
-      default: null,
-    },
-    region: {
-      type: "string",
-      required: false,
-      label: "Province/State",
-      min: FIELD_LIMITS.venueRegion.min,
-      max: FIELD_LIMITS.venueRegion.max,
-      default: null,
-    },
-    postal_code: {
-      type: "string",
-      required: false,
-      label: "Postal code",
-      min: FIELD_LIMITS.venuePostal.min,
-      max: FIELD_LIMITS.venuePostal.max,
-      pattern: POSTAL_CODE_REGEX,
-      patternError: "Postal code must be a valid US ZIP or Canadian postal code",
-      default: null,
-    },
-    country: {
-      type: "string",
-      required: false,
-      label: "Country",
-      min: FIELD_LIMITS.venueCountry.min,
-      max: FIELD_LIMITS.venueCountry.max,
-      default: null,
-    },
-    phone: {
-      type: "string",
-      required: false,
-      label: "Phone",
-      min: FIELD_LIMITS.venuePhone.min,
-      max: FIELD_LIMITS.venuePhone.max,
-      pattern: PHONE_REGEX,
-      patternError: "Phone number must contain only digits and formatting characters",
-      default: null,
-    },
-    contact_email: {
-      type: "email",
-      required: false,
-      label: "Contact email",
-      min: FIELD_LIMITS.venueContactEmail.min,
-      max: FIELD_LIMITS.venueContactEmail.max,
-      default: null,
-    },
-  },
-
-  band: {
-    name: {
-      type: "string",
-      required: true,
-      label: "Band name",
-      min: FIELD_LIMITS.bandName.min,
-      max: FIELD_LIMITS.bandName.max,
-    },
-    eventId: {
-      type: "id",
-      required: false,
-      label: "Event",
-    },
-    venueId: {
-      type: "id",
-      required: false,
-      label: "Venue",
-    },
-    startTime: {
-      type: "time",
-      required: false,
-      label: "Start time",
-    },
-    endTime: {
-      type: "time",
-      required: false,
-      label: "End time",
-    },
-    url: {
-      type: "url",
-      required: false,
-      label: "Website URL",
-      max: FIELD_LIMITS.bandUrl.max,
-    },
-    origin: {
-      type: "string",
-      required: false,
-      label: "Origin",
-      max: FIELD_LIMITS.bandOrigin.max,
-    },
-    origin_city: {
-      type: "string",
-      required: false,
-      label: "Origin city",
-      max: FIELD_LIMITS.bandOriginCity.max,
-    },
-    origin_region: {
-      type: "string",
-      required: false,
-      label: "Origin region",
-      max: FIELD_LIMITS.bandOriginRegion.max,
-    },
-    contact_email: {
-      type: "email",
-      required: false,
-      label: "Contact email",
-      max: FIELD_LIMITS.bandContactEmail.max,
-    },
-    genre: {
-      type: "string",
-      required: false,
-      label: "Genre",
-      max: FIELD_LIMITS.bandGenre.max,
-    },
-    description: {
-      type: "string",
-      required: false,
-      label: "Description",
-      max: FIELD_LIMITS.bandDescription.max,
-    },
-    instagram: {
-      type: "string",
-      required: false,
-      label: "Instagram",
-      max: FIELD_LIMITS.socialHandle.max,
-    },
-    bandcamp: {
-      type: "url",
-      required: false,
-      label: "Bandcamp URL",
-      max: FIELD_LIMITS.bandUrl.max,
-    },
-    facebook: {
-      type: "url",
-      required: false,
-      label: "Facebook URL",
-      max: FIELD_LIMITS.bandUrl.max,
-    },
-  },
-
-  event: {
-    name: {
-      type: "string",
-      required: true,
-      label: "Event name",
-      min: FIELD_LIMITS.eventName.min,
-      max: FIELD_LIMITS.eventName.max,
-    },
-    slug: {
-      type: "string",
-      required: true,
-      label: "Slug",
-      min: FIELD_LIMITS.eventSlug.min,
-      max: FIELD_LIMITS.eventSlug.max,
-      pattern: /^[a-z0-9-]+$/,
-      patternError: "Slug must contain only lowercase letters, numbers, and hyphens",
-    },
-    date: {
-      type: "date",
-      required: true,
-      label: "Date",
-    },
-    end_date: {
-      type: "date",
-      required: false,
-      label: "End date",
-      default: null,
-    },
-    status: {
-      type: "enum",
-      required: false,
-      label: "Status",
-      values: VALID_EVENT_STATUSES,
-      default: "draft",
-    },
-    description: {
-      type: "string",
-      required: false,
-      label: "Description",
-      min: FIELD_LIMITS.eventDescription.min,
-      max: FIELD_LIMITS.eventDescription.max,
-    },
-    city: {
-      type: "string",
-      required: false,
-      label: "City",
-      min: FIELD_LIMITS.eventCity.min,
-      max: FIELD_LIMITS.eventCity.max,
-    },
-    ticket_url: {
-      type: "url",
-      required: false,
-      label: "Ticket link",
-      max: FIELD_LIMITS.ticketLink.max,
-    },
-    poster_url: {
-      type: "url",
-      required: false,
-      label: "Poster image",
-      max: FIELD_LIMITS.eventPosterUrl.max,
-    },
-    venue_info: {
-      type: "string",
-      required: false,
-      label: "Venue info",
-      min: FIELD_LIMITS.eventVenueInfo.min,
-      max: FIELD_LIMITS.eventVenueInfo.max,
-      validate: (value) => {
-        try {
-          JSON.parse(value);
-          return { valid: true };
-        } catch {
-          return { valid: false, error: "Venue info must be valid JSON" };
-        }
-      },
-    },
-    social_links: {
-      type: "string",
-      required: false,
-      label: "Social links",
-      min: FIELD_LIMITS.eventSocialLinks.min,
-      max: FIELD_LIMITS.eventSocialLinks.max,
-      validate: (value) => {
-        try {
-          JSON.parse(value);
-          return { valid: true };
-        } catch {
-          return { valid: false, error: "Social links must be valid JSON" };
-        }
-      },
-    },
-    theme_colors: {
-      type: "string",
-      required: false,
-      label: "Theme colors",
-      min: FIELD_LIMITS.eventThemeColors.min,
-      max: FIELD_LIMITS.eventThemeColors.max,
-      validate: (value) => {
-        try {
-          JSON.parse(value);
-          return { valid: true };
-        } catch {
-          return { valid: false, error: "Theme colors must be valid JSON" };
-        }
-      },
-    },
-    // Schema-level check is deliberately shallow (generic JSON parseability,
-    // mirroring venue_info/social_links above) — the real semantic check
-    // (date-span + HH:MM shape) is `validateDoorsJson`, called explicitly by
-    // the create/update handlers once `date`/`end_date` are known (#569).
-    doors_json: {
-      type: "string",
-      required: false,
-      label: "Doors times",
-      min: FIELD_LIMITS.eventDoorsJson.min,
-      max: FIELD_LIMITS.eventDoorsJson.max,
-      validate: (value) => {
-        try {
-          JSON.parse(value);
-          return { valid: true };
-        } catch {
-          return { valid: false, error: "Doors times must be valid JSON" };
-        }
-      },
-    },
-  },
-
-  user: {
-    email: {
-      type: "email",
-      required: true,
-      label: "Email",
-    },
-    password: {
-      type: "string",
-      required: true,
-      label: "Password",
-      min: FIELD_LIMITS.password.min,
-      max: FIELD_LIMITS.password.max,
-      validate: (value) => {
-        const result = validatePassword(value, {
-          minLength: FIELD_LIMITS.password.min,
-          requireUppercase: true,
-          requireLowercase: true,
-          requireNumber: true,
-          requireSpecial: true,
-        });
-        return result.valid ? { valid: true } : { valid: false, error: result.errors[0] };
-      },
-    },
-    firstName: {
-      type: "string",
-      required: true,
-      label: "First name",
-      min: FIELD_LIMITS.userFirstName.min,
-      max: FIELD_LIMITS.userFirstName.max,
-    },
-    lastName: {
-      type: "string",
-      required: true,
-      label: "Last name",
-      min: FIELD_LIMITS.userLastName.min,
-      max: FIELD_LIMITS.userLastName.max,
-    },
-    role: {
-      type: "enum",
-      required: true,
-      label: "Role",
-      values: VALID_ROLES,
-    },
-  },
-  userInvite: {
-    email: {
-      type: "email",
-      required: true,
-      label: "Email",
-    },
-    firstName: {
-      type: "string",
-      required: true,
-      label: "First name",
-      min: FIELD_LIMITS.userFirstName.min,
-      max: FIELD_LIMITS.userFirstName.max,
-    },
-    lastName: {
-      type: "string",
-      required: true,
-      label: "Last name",
-      min: FIELD_LIMITS.userLastName.min,
-      max: FIELD_LIMITS.userLastName.max,
-    },
-    role: {
-      type: "enum",
-      required: true,
-      label: "Role",
-      values: VALID_ROLES,
-    },
-  },
-
-  inviteCode: {
-    role: {
-      type: "enum",
-      required: true,
-      label: "Role",
-      values: VALID_ROLES,
-    },
-    expiryDays: {
-      type: "number",
-      required: false,
-      label: "Expiry days",
-      min: 1,
-      max: 365,
-      default: 7,
-    },
-    restrictedEmail: {
-      type: "email",
-      required: false,
-      label: "Restricted email",
-    },
-  },
-};
+// Input validation utilities for Cloudflare Workers — facade module (#906).
+//
+// This file used to hold ~1,460 lines and 31 exports directly, and was
+// imported by 43 files across functions/. That made it a single point of
+// contention for unrelated changes and hard to navigate. The implementation
+// has been split into focused sibling modules under
+// functions/utils/validation/ (fieldLimits, identity, strings, urls,
+// contact, datetime, ids, schema, response), grouped by what they validate.
+//
+// This file is now a pure re-exporting facade: every name it used to export
+// directly, it now re-exports from the module that owns it. Every existing
+// `import { X } from "…/utils/validation"` (this file's own path) across the
+// codebase keeps working unchanged — nothing outside functions/utils/ should
+// ever need to change its import path because of this split.
+//
+// functions/utils/__tests__/validationPublicSurface.test.js pins the full
+// list of 31 export names and their types — run it after touching this file
+// or any module it re-exports from.
+
+export { FIELD_LIMITS } from "./validation/fieldLimits.js";
+export { isValidEmail, validatePassword, isValidUUID, isValidRole } from "./validation/identity.js";
+export { sanitizeString, sanitizeOptionalText, validateRequiredFields } from "./validation/strings.js";
+export {
+  normalizeHttpUrl,
+  safeReflectHandleOrUrl,
+  safeReflectSocialLinks,
+  safeReflectSocialLinksString,
+  sanitizeOptionalHttpUrl,
+  sanitizeBandSocialLinks,
+  sanitizeEventSocialLinks,
+  sanitizeVenueInfo,
+  isValidURL,
+} from "./validation/urls.js";
+export { isValidPhone, isValidPostalCode, normalizePostalCode } from "./validation/contact.js";
+export {
+  isValidISODate,
+  isValidTime,
+  validateSetTimes,
+  validateDate,
+  validatePerformanceDate,
+  validateDoorsJson,
+} from "./validation/datetime.js";
+export { validateId, validateIdArray } from "./validation/ids.js";
+export { validateEntity, VALIDATION_SCHEMAS } from "./validation/schema.js";
+export { validationErrorResponse } from "./validation/response.js";
